requirements.txt
selenium==4.9.0
requests==2.28.1
matplotlib==3.5.2
pymongo==4.3.3
pillow==9.2.0